Fix tearDown method to check whether a repository exists before trying to delete it.
* tests/svn_fs.py
  (tearDown): Check whether a repository exists before trying to delete it.
  (setUp): Don't catch exceptions from tearDown, because tearDown now
  succeeds when the repository is missing.

Suggested by: dlr
Thread: http://svn.haxx.se/dev/archive-2005-09/1071.shtml



git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@856393 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/swig/python/tests/trac/versioncontrol/tests/svn_fs.py
@@ -44,10 +44,7 @@ class SubversionRepositoryTestSetup(TestSetup):
 
         # Remove the trac-svnrepos directory, so that we can
         # ensure a fresh start.
-        try:
-          self.tearDown()
-        except:
-          pass
+        self.tearDown()
 
         try:
             r = repos.svn_repos_create(REPOS_PATH, '', '', None, None, pool)
@@ -67,14 +64,18 @@ class SubversionRepositoryTestSetup(TestSetup):
             core.apr_terminate()
 
     def tearDown(self):
-        if os.name == 'nt':
-            # The Windows version of 'shutil.rmtree' doesn't override the
-            # permissions of read-only files, so we have to do it ourselves:
-            format_file = os.path.join(REPOS_PATH, 'db', 'format')
-            if os.path.isfile(format_file):
-                os.chmod(format_file, stat.S_IRWXU)
-            os.chmod(os.path.join(REPOS_PATH, 'format'), stat.S_IRWXU)
-        shutil.rmtree(REPOS_PATH)
+        if os.path.exists(REPOS_PATH):
+            if os.name == 'nt':
+                # The Windows version of 'shutil.rmtree' doesn't override the
+                # permissions of read-only files, so we have to do it
+                # ourselves:
+                db_format_file = os.path.join(REPOS_PATH, 'db', 'format')
+                if os.path.isfile(db_format_file):
+                    os.chmod(db_format_file, stat.S_IRWXU)
+                format_file = os.path.join(REPOS_PATH, 'format')
+                if os.path.isfile(format_file):
+                    os.chmod(format_file, stat.S_IRWXU)
+            shutil.rmtree(REPOS_PATH)
 
 
 class SubversionRepositoryTestCase(unittest.TestCase):
